hospital-erp/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0
pytz==2023.3.post1
